pyproject.toml
[project]
name = "tortoise-orm-migrate-timezones-bug-reporoduce"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.5",
    "tortoise-orm[accel,asyncmy]",
]

[tool.tortoise]
tortoise_orm = "main.TORTOISE_ORM"

[tool.uv.sources]
tortoise-orm = { git = "https://github.com/tortoise/tortoise-orm.git", rev = "develop" }
